bank transfer
bank transfer payment method added

+++ database/migrations/2022_01_16_153459_create_items_table.php
@@ -21,7 +21,7 @@ class CreateItemsTable extends Migration
             $table->string('photo');
             $table->string('thumb');
             $table->string('color')->nullable();
-            $table->double('price');
+            $table->float('price', 10, 2);
             $table->string('width')->nullable();
             $table->string('height')->nullable();
             $table->string('diameter')->nullable();
