+++ database/migrations/2022_01_30_183509_create_shipping_methods_table.php
@@ -18,7 +18,8 @@ class CreateShippingMethodsTable extends Migration
             $table->string('name');
             $table->string('slug');
             $table->longText('detail');
-            $table->double('price');
+
+            $table->float('price', 10, 2);
             $table->string('thumb');
             $table->timestamps();
         });
